增加在线人数逻辑

+++ src/main/java/com/ruoyi/system/domain/User.java
@@ -181,6 +181,6 @@ public class User {
 
     private Date endTime;
 
-    private List<Long> onlineUserIdList;
+    private List<Integer> onlineUserIdList;
 
 }
